BackEnd/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
Django==4.2.14
elastic-transport==8.15.1
elasticsearch==8.7.0
elasticsearch-dsl==8.16.0
executing==2.1.0
idna==3.7
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.2
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
zipp==3.20.2
redis==5.2.1
django-redis==5.4.0
celery==5.4.0
mysql-connector-python=9.1.0